修改所有编码的数据类型为int

+++ src/main/java/com/guotu/gt/controller/BasicinfoLoginLogDTOController.java
@@ -23,7 +23,7 @@ public class BasicinfoLoginLogDTOController {
 
     @DeleteMapping
     @ApiOperation("根据编码删除一条登陆日志")
-    public Result<Object> delete(@RequestParam("code") Byte code){
+    public Result<Object> delete(@RequestParam("code") int code){
         basicinfoLoginLogDTOService.delete(code);
         return ResultUtil.success();
     }
